refactor: move attention pooling to coca
